refactor: combine `question_answering` and `query_llm`
